Refactor: 에러코드 통합
